refactor: 데일리 FSD에 맞게 정리 (#145)
