+++ src/features/community-quiz/model/useAnswerQuizMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query'
 import { toast } from 'sonner'
-import { answerQuizAction } from '@/widgets/community-quiz/api/answerQuizAction'
+import { answerQuizAction } from '@/features/community-quiz/api/answerQuizAction'
 import { QuizAnswerForm } from '@/entities/quiz/model/schema'
 
 export function useAnswerQuizMutation() {
