--- src/features/community/api/query-keys.ts
@@ -1,5 +1,0 @@
-export const queryKeys = {
-  all: ['community'] as const,
-  quiz: () => [...queryKeys.all, 'quiz'] as const,
-  quizResult: () => [...queryKeys.quiz(), 'result'] as const,
-}
