--- src/features/community/assets/index.ts
@@ -1,6 +1,0 @@
-import QuizImage from './quiz-image.png'
-import QuizImageWide from './quiz-image-wide.png'
-import QuoteImage from './quote-image.png'
-import QuoteImageWide from './quote-image-wide.png'
-
-export { QuizImage, QuizImageWide, QuoteImage, QuoteImageWide }
