--- src/features/community/ui/index.ts
@@ -1,3 +1,0 @@
-// import CommunityBanner from '@/widgets/community-quiz/ui/CommunityBanner'
-
-// export { CommunityBanner }
