+++ src/widgets/community-quiz/assets/quiz-image-wide.png
