+++ src/widgets/community-quiz/assets/quiz-image.png
