+++ src/widgets/community-quiz/assets/quote-image-wide.png
