+++ src/widgets/community-quiz/assets/quote-image.png
